+++ uitools/register/Esri/ArcGISRuntime/Toolkit/internal/register_cpp.cpp
@@ -19,6 +19,8 @@
 #include "AuthenticationController.h"
 #include "BasemapGalleryController.h"
 #include "BasemapGalleryItem.h"
+#include "BookmarksViewController.h"
+#include "BookmarkListItem.h"
 #include "CoordinateConversionController.h"
 #include "CoordinateConversionOption.h"
 #include "CoordinateConversionResult.h"
@@ -168,6 +170,8 @@ namespace Toolkit {
     registerComponent<AuthenticationController>(10);
     registerComponent<BasemapGalleryController>(12);
     registerComponent<BasemapGalleryItem>(12);
+    registerComponent<BookmarksViewController>(15);
+    registerComponent<BookmarkListItem>(15);
     registerComponent<CoordinateConversionController>(10);
     registerComponent<CoordinateConversionOption>(10);
     registerComponent<CoordinateConversionResult>(10);